Update WPI
CS 2102 d20 lectures are freely available so switch to d18 with d20 lectures since d18 is the last d term with materials and d20 only has lectures. Switch to CS 1101 to c18 so it's the one before CS 2102. Switch CS 2223, which has no lectures, to Princeton Algorithms, which uses the same textbook.