Remove Algorithms and MIT Calculus
Decided to set the scope for an Associate of Science in Computer Science so Algorithms is gone. Fall 2018 is the earliest Berkeley CS 170 that has lectures post-Fall 2014 for future reference in case that's wanted again.

Removed MIT Calculus it expects too much coming in for the average person to take it without actually refreshing on PreCalculus, but then it doesn't even teach the material all that well. Sydney's course has great reviews and you don't need to have taken PreCalculus recently and then UPenn really goes into Calculus on another level with teaching "Calculus II". I feel the sequence serves people much better than the MIT course. Also, Sydney's course is pretty similar to Berkeley's Math 1A, which is a corequisite to CS 61A.