Add PC Architecture
Berkeley's course is the course of choice because no sequence teaches C prior to this. Brown uses Carnegie Mellon's course because that's the sequence for people already programming a bunch so they're given the responsibility of picking up a new language despite the fact that everyone should be able to do that by then.